Move override_fedora.py to _internal
