+++ certbot-apache/certbot_apache/entrypoint.py
@@ -7,7 +7,7 @@ from certbot import util
 
 from certbot_apache import configurator
 from certbot_apache import override_arch
-from certbot_apache import override_fedora
+from certbot_apache._internal import override_fedora
 from certbot_apache._internal import override_darwin
 from certbot_apache._internal import override_debian
 from certbot_apache._internal import override_centos
